html/obsolete/requirements-for-implementations/other-elements-attributes-and-apis/document-color-fg.html

<!DOCTYPE html>
<title>document wg color</title>
<meta charset=utf-8>
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-fgcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-linkcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-vlinkcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-alinkcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html#dom-document-bgcolor">
<link rel="help" href="https://html.spec.whatwg.org/multipage/obsolete.html">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<frameset>
  <frame src = "body-container.html">
</frameset>
<iframe id="test" src="body-container.html"></iframe>
<script>
var doc=document.getElementById("test");
doc.body.parentNode.removeChild(doc.body);

test(function() {
  assert_equals(doc.body, null);
}, "body is gone")
// Getting Part
test(function() {
assert_equals(doc.fgColor,"");
assert_equals(doc.bgColor,"");
assert_equals(doc.linkColor,"");
assert_equals(doc.vlinkColor,"");
assert_equals(doc.alinkColor,"");

}, "getting the document color attributes with a frameset")

// Setting Part
// TODO

</script>
